Changed sample set to one minute's worth, changed requests to requests Per Second
