+++ src/server/status/index.js
@@ -1,24 +1,12 @@
 module.exports = function (kbnServer, server, config) {
   var _ = require('lodash');
-  var Samples = require('./Samples');
   var ServerStatus = require('./ServerStatus');
   var { join } = require('path');
 
   kbnServer.status = new ServerStatus(kbnServer.server);
-  kbnServer.metrics = new Samples(1);
 
   if (server.plugins.good) {
-    server.plugins.good.monitor.on('ops', function (event) {
-      var port = config.get('server.port');
-      kbnServer.metrics.add({
-        heapTotal: event.psmem.heapTotal,
-        heapUsed: event.psmem.heapUsed,
-        load: event.osload,
-        responseTimeAvg: _.get(event, ['responseTimes', port, 'avg']),
-        responseTimeMax: _.get(event, ['responseTimes', port, 'max']),
-        requests: _.get(event, ['requests', port, 'total'], 0)
-      });
-    });
+    kbnServer.mixin(require('./metrics'));
   }
 
   server.route({
